Implement an approach for long-running custom resource functions
If a resource function (`create`/`update`/`delete`) returns an object
with a `continuationAfter` member, then instead of writing the response
we set up a CloudWatch Events rule and target to invoke the function again
after the given number of seconds (rounded to the full minute)

The `continuationAttributes` also provided in the result will be provided
for that invocation in a separate parameter (to not mix up the schema more
than needed), so that the next invocation can continue where it left off.

The main idea here is that this is reasonably easy to work with (compared
to cfn-lambda's API), and also should be cheaper as we do not need to
"busy-wait" by invoking the function continuously.

The first user will be the ACM certificate function.