Document that the user-pool-domain-attributes resource should be given the UserPoolId
The user pool domain itself is a name that doesn't change when the domain is moved from one
user pool to another, but the attributes would change in that case (a new CloudFront distribution
will be created in the new user pool).

This is basically the flip-side of the problem that user pool domains are a PITA to move, because
they link two very different things together.