Fix the logic for bumping the minute up to avoid scheduling problems
We cannot just add 1 to the minutes, as that might give us "60" at the end
of the hour -- and then we would need to bump the hours, too. Instead change
the calculation so that we check the condition explicitly ("target minute must be
strictly bigger than current minute"), and then use `Date.prototype.setMinutes`, which
will handle the overflow for us.